financial_data_collector/alpaca_data_collector/pyproject.toml
[tool.poetry]
name = "alpaca-data-collector"
version = "1.0.0"
description = "Alpaca Markets API data collector for the AI Trading Platform"
authors = ["Trading Platform Team"]

[tool.poetry.dependencies]
python = "^3.11"
alpaca-trade-api = "^3.1.1"
pandas = "^2.0.0"
pyyaml = "^6.0"
jsonschema = "^4.17.0"
pyarrow = "^12.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

